+++ src/broker/mod.rs
@@ -14,7 +14,10 @@ use crate::{
 };
 
 mod amqp;
+mod redis;
+pub use self::redis::{RedisBroker, RedisBrokerBuilder};
 pub use amqp::{AMQPBroker, AMQPBrokerBuilder};
+
 #[cfg(test)]
 pub mod mock;